Refactor: util을 외부모듈로 하여 하나로 통합
