+++ app.js
@@ -1,6 +1,6 @@
 import { makeCarousel } from './src/components/jungleUI.js';
 import { renderCard } from './src/template.js';
-import { getData } from './src/util.js';
+import { getData } from './modules/JinUtil/index.js';
 
 const CARD_URL = './resources/localData.json';
