+++ modules/JinUtil/index.js
@@ -18,3 +18,5 @@ export const isContainClass = (target, className) => target.classList.contains(c
 export const sleep = delay => {
   return new Promise(resolve => setTimeout(resolve, delay));
 };
+
+export const getData = url => fetch(url).then(response => response.json());
