--- src/util.js
@@ -1,1 +1,0 @@
-export const getData = url => fetch(url).then(response => response.json());
